+++ visualization/app/codeCharta/state/store/appSettings/invertArea/invertArea.splitter.ts
@@ -1,0 +1,5 @@
+import { setInvertArea } from "./invertArea.actions"
+
+export function splitInvertAreaAction(payload: boolean) {
+	return setInvertArea(payload)
+}