+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased] (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
+### Added 🚀
+
+-   add invert the amount of area, which is used for the nodes [#697](https://github.com/MaibornWolff/codecharta/issues/697)
+
 ## [1.88.0] - 2022-01-24
 
 ### Changed